# perf: Don't open the output file each time we write.