This improves performance of the the file output mode, which should
make it easier to improve the performance of the core generation
logic.

Benchmarked via:

```
time \
./target/release/iox_data_generator \
--spec iox_data_generator/schemas/fully-supported.toml \
--output /tmp/out \
--start '1 month ago'
```

Before:

```
Submitted 271608 total points

real	10.912	10911567us
user	3.129	3129032us
sys	6.257	6257340us
cpu	86%
mem	7152 KiB
```

After:

```
Submitted 271588 total points

real	2.291	2291364us
user	1.969	1969357us
sys	0.058	58030us
cpu	88%
mem	7104 KiB
```

That's 21.0% of the previous time.